+++ release_notes.md
@@ -3,3 +3,6 @@
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
+-->
+- Log enum names instead of integers in options formatting for `TimerTriggerPlatformOptions` (#11689)
+- Update WebHostWorkerRuntimeResolver to always read from IConfiguration variable value for FWR (#11720)